// src/interface/layout/tagsView.interface.ts
export type TagItem = {
  code: string;

  label:  string;
  /** tag's route path */
  path: string;

  /** can be closed ? */
  closable: boolean;
};

export interface TagState {
  /** tagsView list */
  tags: TagItem[];

  /**current tagView id */
  activeTagId: TagItem['path'];
}


export interface Tag {
  path: string;
  label: string ;
}